Hardening of ConDepNode file-handling, and added logging to client.
- Added detection and logging of errors when overwriting files on server.
If a file is readonly on server when deploying the file-copy stops silently
and the user gets no notification.

- Exception thrown in ExecutePostProcessingAsync in MultipartSyncDirStreamProvider
deadlocks the controller and ConDep times out without the user getting any info.
A nasty bug in the framework makes this happen about 1% of the time when setting
attributes and LastWriteTime.
(see http://aspnetwebstack.codeplex.com/workitem/282)

- Logs errors on client side for so user can do something about the error.